cookie-store/cookieStore_getAll_domain_ordering_unprefixed.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>Unprefixed cookie getAll domain ordering</title>
<meta name="help"
      href="https://github.com/WICG/cookie-store/blob/gh-pages/explainer.md">
<meta name="help"
      href="https://tools.ietf.org/html/draft-ietf-httpbis-rfc6265bis-02#section-4.1.2.3">
<meta name="help"
      href="https://tools.ietf.org/html/draft-ietf-httpbis-rfc6265bis-02#section-5.5">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-helpers.js"></script>
<script>
'use strict';

const kMyDomain = document.domain;

promise_test(async testCase => {
  try {
    await cookieStore.set(
        'unprefixed_getAll_domain_shadowing1_unshadowed',
        'value1_implicit_domain_' + kMyDomain);
    await cookieStore.set(
        'unprefixed_getAll_domain_shadowing2_shadowed',
        'value2_implicit_domain_' + kMyDomain);
    await cookieStore.set(
        'unprefixed_getAll_domain_shadowing3_unshadowed',
        'value3_implicit_domain_' + kMyDomain);
    await cookieStore.set(
        'unprefixed_getAll_domain_shadowing2_shadowed',
        'value2_explicit_domain_' + kMyDomain,
        {
          domain: kMyDomain,
        });
    assert_equals(
        cookieString(await cookieStore.getAll()),
        [
          'unprefixed_getAll_domain_shadowing2_shadowed=' +
              'value2_explicit_domain_' + kMyDomain,
          'unprefixed_getAll_domain_shadowing1_unshadowed=' +
              'value1_implicit_domain_' + kMyDomain,
          'unprefixed_getAll_domain_shadowing2_shadowed=' +
              'value2_implicit_domain_' + kMyDomain,
          'unprefixed_getAll_domain_shadowing3_unshadowed=' +
              'value3_implicit_domain_' + kMyDomain,
        ].join('; '));
  } finally {
    await cookieStore.delete('unprefixed_getAll_domain_shadowing1_unshadowed');
    await cookieStore.delete('unprefixed_getAll_domain_shadowing2_shadowed');
    await cookieStore.delete('unprefixed_getAll_domain_shadowing3_unshadowed');
    await cookieStore.delete('unprefixed_getAll_domain_shadowing2_shadowed', {
      domain: kMyDomain,
    });
  }
}, 'getAll item order with domain shadowing and unshadowed cookies');

</script>
